+++ ureport/assets/tests.py
@@ -1,5 +1,5 @@
 from django.conf import settings
-from django.core.urlresolvers import reverse
+from django.urls import reverse
 from ureport.assets.models import Image
 from ureport.tests import UreportTest