+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+v1.0.344
+----------
+ * Use big int for poll results counter primary key, migrations
+ * Update dependencies
+
+v1.0.343
+----------
+ * Merge pull request #293 from rapidpro/result-text-field
+
+v1.0.341
+----------
+ * More italian
+ * Display only featured stories on home page
+
 v1.0.340
 ----------
  * Update Italian